refactor(utils/misc): separte out number-related utils and more add docstring to the functions in `utils/numbers` also, move `getDiscordAvatarUrl` back into misc because it was causing mt to not open somehow for #5187
