+++ frontend/src/ts/elements/account-button.ts
@@ -3,7 +3,6 @@ import { isAuthenticated } from "../firebase";
 import * as Misc from "../utils/misc";
 import { getAll } from "./theme-colors";
 import * as SlowTimer from "../states/slow-timer";
-import { getDiscordAvatarUrl } from "../utils/url-handler";
 
 let usingAvatar = false;
 
@@ -109,7 +108,7 @@ export async function update(
       });
     }
     if ((discordAvatar ?? "") && (discordId ?? "")) {
-      void getDiscordAvatarUrl(discordId, discordAvatar).then(
+      void Misc.getDiscordAvatarUrl(discordId, discordAvatar).then(
         (discordAvatarUrl) => {
           if (discordAvatarUrl !== null) {
             $("header nav .account .avatar").css(
