+++ frontend/src/ts/test/result.ts
@@ -13,6 +13,7 @@ import * as QuoteRatePopup from "../popups/quote-rate-popup";
 import * as GlarsesMode from "../states/glarses-mode";
 import * as SlowTimer from "../states/slow-timer";
 import * as Misc from "../utils/misc";
+import * as Numbers from "../utils/numbers";
 import { get as getTypingSpeedUnit } from "../utils/typing-speed-units";
 import * as FunboxList from "./funbox/funbox-list";
 import * as PbCrown from "./pb-crown";
@@ -53,7 +54,7 @@ async function updateGraph(): Promise<void> {
 
   for (let i = 1; i <= TestInput.wpmHistory.length; i++) {
     if (TestStats.lastSecondNotRound && i === TestInput.wpmHistory.length) {
-      labels.push(Misc.roundTo2(result.testDuration).toString());
+      labels.push(Numbers.roundTo2(result.testDuration).toString());
     } else {
       labels.push(i.toString());
     }
@@ -64,7 +65,7 @@ async function updateGraph(): Promise<void> {
 
   const chartData1 = [
     ...TestInput.wpmHistory.map((a) =>
-      Misc.roundTo2(typingSpeedUnit.fromWpm(a))
+      Numbers.roundTo2(typingSpeedUnit.fromWpm(a))
     ),
   ];
 
@@ -72,7 +73,7 @@ async function updateGraph(): Promise<void> {
 
   const chartData2 = [
     ...result.chartData.raw.map((a) =>
-      Misc.roundTo2(typingSpeedUnit.fromWpm(a))
+      Numbers.roundTo2(typingSpeedUnit.fromWpm(a))
     ),
   ];
 
@@ -88,7 +89,7 @@ async function updateGraph(): Promise<void> {
 
   let smoothedRawData = chartData2;
   if (!useUnsmoothedRaw) {
-    smoothedRawData = Misc.smooth(smoothedRawData, 1);
+    smoothedRawData = Numbers.smooth(smoothedRawData, 1);
     smoothedRawData = smoothedRawData.map((a) => Math.round(a));
   }
 
@@ -175,7 +176,7 @@ export async function updateGraphPBLine(): Promise<void> {
   );
   if (lpb === 0) return;
   const typingSpeedUnit = getTypingSpeedUnit(Config.typingSpeedUnit);
-  const chartlpb = Misc.roundTo2(typingSpeedUnit.fromWpm(lpb)).toFixed(2);
+  const chartlpb = Numbers.roundTo2(typingSpeedUnit.fromWpm(lpb)).toFixed(2);
   resultAnnotation.push({
     display: true,
     type: "line",
@@ -247,9 +248,9 @@ function updateWpmAndAcc(): void {
       $("#result .stats .raw .bottom").removeAttr("aria-label");
     }
 
-    let time = Misc.roundTo2(result.testDuration).toFixed(2) + "s";
+    let time = Numbers.roundTo2(result.testDuration).toFixed(2) + "s";
     if (result.testDuration > 61) {
-      time = Misc.secondsToString(Misc.roundTo2(result.testDuration));
+      time = Misc.secondsToString(Numbers.roundTo2(result.testDuration));
     }
     $("#result .stats .time .bottom .text").text(time);
     // $("#result .stats .acc .bottom").removeAttr("aria-label");
@@ -309,7 +310,7 @@ function updateConsistency(): void {
 }
 
 function updateTime(): void {
-  const afkSecondsPercent = Misc.roundTo2(
+  const afkSecondsPercent = Numbers.roundTo2(
     (result.afkDuration / result.testDuration) * 100
   );
   $("#result .stats .time .bottom .afk").text("");
@@ -322,9 +323,9 @@ function updateTime(): void {
   );
 
   if (Config.alwaysShowDecimalPlaces) {
-    let time = Misc.roundTo2(result.testDuration).toFixed(2) + "s";
+    let time = Numbers.roundTo2(result.testDuration).toFixed(2) + "s";
     if (result.testDuration > 61) {
-      time = Misc.secondsToString(Misc.roundTo2(result.testDuration));
+      time = Misc.secondsToString(Numbers.roundTo2(result.testDuration));
     }
     $("#result .stats .time .bottom .text").text(time);
   } else {
@@ -335,7 +336,7 @@ function updateTime(): void {
     $("#result .stats .time .bottom .text").text(time);
     $("#result .stats .time .bottom").attr(
       "aria-label",
-      `${Misc.roundTo2(result.testDuration)}s (${
+      `${Numbers.roundTo2(result.testDuration)}s (${
         result.afkDuration
       }s afk ${afkSecondsPercent}%)`
     );
@@ -497,7 +498,7 @@ async function updateTags(dontSave: boolean): Promise<void> {
         ).removeClass("hidden");
         $(`#result .stats .tags .bottom div[tagid="${tag._id}"]`).attr(
           "aria-label",
-          "+" + Misc.roundTo2(result.wpm - tpb)
+          "+" + Numbers.roundTo2(result.wpm - tpb)
         );
         // console.log("new pb for tag " + tag.display);
       } else {
@@ -526,7 +527,7 @@ async function updateTags(dontSave: boolean): Promise<void> {
             position: annotationSide,
             xAdjust: labelAdjust,
             enabled: true,
-            content: `${tag.display} PB: ${Misc.roundTo2(
+            content: `${tag.display} PB: ${Numbers.roundTo2(
               typingSpeedUnit.fromWpm(tpb)
             ).toFixed(2)}`,
           },
